refactor(common): 닉네임 정책과 회원 검증 책임 분리
공통에는 인증·회원이 함께 쓰는 닉네임 정책만 두고, Jakarta 제약조건과 Validator는 회원 모듈 내부로 이동한다. 정책 테스트와 Validator 테스트도 패키지 책임에 맞게 분리한다.

Tested: ./gradlew compileJava; 닉네임 정책·Validator 테스트; AuthApiTest·MemberApiTest·PublicMemberApiTest; CheckmoApplicationTests

Not-tested: 전체 테스트는 기존 Apple JWT 만료 3건과 Sentry workflow 1건 실패
