+++ src/main/java/checkmo/member/internal/service/command/MemberFollowCommandService.java
@@ -1,6 +1,6 @@
 package checkmo.member.internal.service.command;
 
-import checkmo.common.validation.NicknamePolicy;
+import checkmo.common.nickname.NicknamePolicy;
 import checkmo.member.MemberEvent;
 import checkmo.member.internal.entity.Follow;
 import checkmo.member.internal.entity.Member;
